requirements.txt
annotated-types==0.7.0
anyio==4.7.0
Arpeggio==2.0.2
arrow==1.3.0
asttokens==3.0.0
asyncclick==8.1.7.2
attrs==24.3.0
binaryornot==0.4.4
boto3==1.35.87
botocore==1.35.87
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.8
cookiecutter==2.6.0
debugpy==1.8.11
decorator==5.1.1
dill==0.3.9
executing==2.1.0
fastapi==0.115.6
grpcio==1.66.2
h11==0.14.0
idna==3.10
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
-e git+https://github.com/bglick13/next-data.git@ab703b0801f3669d09fea2f184bbbd75b3423d3c#egg=nextdata
parso==0.8.4
parver==0.5
pexpect==4.9.0
prompt_toolkit==3.0.48
protobuf==4.25.5
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pulumi==3.144.1
pulumi_aws==6.66.0
pure_eval==0.2.3
py4j==0.10.9.7
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
pyspark==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
s3transfer==0.10.4
semver==2.13.0
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
text-unidecode==1.3
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
wcwidth==0.2.13
wheel==0.44.0
